1.增加首页 2.增加主机配置和数据库配置 3.数据库连接全通过database.properties配置

+++ src/main/java/tpf/dao/UsersPOMapper.java
@@ -3,13 +3,13 @@ package tpf.dao;
 import tpf.pojo.UsersPO;
 
 public interface UsersPOMapper {
-    int deleteByPrimaryKey(Integer userId);
+    int deleteByPrimaryKey(Long userId);
 
     int insert(UsersPO record);
 
     int insertSelective(UsersPO record);
 
-    UsersPO selectByPrimaryKey(Integer userId);
+    UsersPO selectByPrimaryKey(Long userId);
 
     int updateByPrimaryKeySelective(UsersPO record);
